# Preallocate response buffer in C++ HTTP client