Preallocating response buffer for large responses improves the client receive time considerably (**20683 usec** to **595 usec**). Experiments were performed on `resnet50_fp16_plan` :

## **Before**

### **V1**

```
*** Measurement Settings ***
  Batch size: 128
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 921.6 infer/sec. Avg latency: 137855 usec (std 5894 usec)
  Pass [2] throughput: 896 infer/sec. Avg latency: 141616 usec (std 5400 usec)
  Pass [3] throughput: 947.2 infer/sec. Avg latency: 136513 usec (std 4606 usec)
  Client: 
    Request count: 37
    Throughput: 947.2 infer/sec
    Avg latency: 136513 usec (standard deviation 4606 usec)
    p50 latency: 138071 usec
    p90 latency: 141941 usec
    p95 latency: 142441 usec
    p99 latency: 144178 usec
    Avg HTTP time: 136671 usec (send 78720 usec + response wait 51958 usec + **receive 5993 usec**)
  Server: 
    Request count: 44
    Avg request latency: 41200 usec (overhead 12 usec + queue 68 usec + compute 41120 usec)


```
### **V2**

```
*** Measurement Settings ***
  Batch size: 128
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 768 infer/sec. Avg latency: 162286 usec (std 1634 usec)
  Pass [2] throughput: 793.6 infer/sec. Avg latency: 161110 usec (std 2759 usec)
  Pass [3] throughput: 768 infer/sec. Avg latency: 162715 usec (std 2264 usec)
  Client: 
    Request count: 30
    Throughput: 768 infer/sec
    Avg latency: 162715 usec (standard deviation 2264 usec)
    p50 latency: 162237 usec
    p90 latency: 166181 usec
    p95 latency: 166958 usec
    p99 latency: 167194 usec
    Avg HTTP time: 162531 usec (send 89773 usec + response wait 52075 usec + **receive 20683 usec**)
  Server: 
    Inference count: 4736
    Execution count: 37
    Successful request count: 37
    Avg request latency: 41033 usec (overhead 2 usec + queue 66 usec + compute input 19119 usec + compute infer 21731 usec + compute output 115 usec)

Inferences/Second vs. Client Average Batch Latency
Concurrency: 1, throughput: 768 infer/sec, latency 162715 usec
```

## **After**

### **V2**

```
*** Measurement Settings ***
  Batch size: 128
  Measurement window: 5000 msec
  Using synchronous calls for inference
  Stabilizing using average latency

Request concurrency: 1
  Pass [1] throughput: 870.4 infer/sec. Avg latency: 144043 usec (std 2676 usec)
  Pass [2] throughput: 896 infer/sec. Avg latency: 141958 usec (std 3738 usec)
  Pass [3] throughput: 921.6 infer/sec. Avg latency: 142499 usec (std 3419 usec)
  Client: 
    Request count: 36
    Throughput: 921.6 infer/sec
    Avg latency: 142499 usec (standard deviation 3419 usec)
    p50 latency: 142148 usec
    p90 latency: 148352 usec
    p95 latency: 148756 usec
    p99 latency: 152596 usec
    Avg HTTP time: 142344 usec (send 89756 usec + response wait 51993 usec + receive **595 usec**)
  Server: 
    Inference count: 5504
    Execution count: 43
    Successful request count: 43
    Avg request latency: 41292 usec (overhead 3 usec + queue 65 usec + compute input 19352 usec + compute infer 21762 usec + compute output 110 usec)
```