@GuanLuo 2020-08-04
Good catch. Now the ratio of send / recv time matches the ratio of input / output size